Remove commented and unnecessary code, better query context
